# backend/src/openapi.yaml
openapi: 3.0.3
info:
  title: Momento API
  description: 프로젝트 관리 및 스터디 세션 추적 앱 API
  version: 1.0.0
  contact:
    name: Momento Team
servers:
  - url: http://localhost:3000/api
    description: Development
  - url: http://172.10.5.61:3000/api
    description: Production
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID Token
      description: Firebase Authentication ID Token
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firebaseUid:
          type: string
        nickname:
          type: string
        profileEmoji:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserSearchResult:
      type: object
      properties:
        id:
          type: string
          format: uuid
        nickname:
          type: string
        profileEmoji:
          type: string
    CreateUserDto:
      type: object
      required:
        - nickname
      properties:
        nickname:
          type: string
          example: 사용자닉네임
        profileEmoji:
          type: string
          example: "😀"
    UpdateUserDto:
      type: object
      properties:
        nickname:
          type: string
        profileEmoji:
          type: string
...
    TimeLog:
      type: object
      properties:
        id:
          type: string
          format: uuid
        checklistId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
          nullable: true
        durationMinutes:
          type: integer
          nullable: true
...
    Receipt:
      type: object
      properties:
        id:
          type: string
          format: uuid
        date:
          type: string
          format: date
        imageUrl:
          type: string
        totalMinutes:
          type: integer
        completedTasksCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    CreateReceiptDto:
      type: object
      required:
        - date
        - imageUrl
      properties:
        date:
          type: string
          format: date
          example: "2025-01-10"
        imageUrl:
          type: string
    PaginatedResponse:
      type: object
      properties:
        data:
          type: array
          items: {}
        meta:
          type: object
          properties:
            total:
              type: integer
            page:
              type: integer
            limit:
              type: integer
...
paths:
...
  /users/me:
    get:
      tags: [Users]
      summary: 내 정보 조회
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: 사용자를 찾을 수 없습니다

    patch:
      tags: [Users]
      summary: 내 정보 수정
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserDto'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users:
    post:
      tags: [Users]
      summary: 회원가입 (최초 로그인 시)
      description: firebaseUid는 Authorization 헤더의 Firebase ID Token에서 서버가 자동 추출합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserDto'
      responses:
        '201':
          description: 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users/search:
    get:
      tags: [Users]
      summary: 닉네임으로 사용자 검색
      description: UI 자동완성/실시간 검증용
      parameters:
        - name: nickname
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSearchResult'
        '404':
          description: 사용자를 찾을 수 없습니다

  # ===== Projects =====
...
  /checklists/{checklistId}/time-logs/start:
    post:
      tags: [Time Logs]
      summary: 타이머 시작
      parameters:
        - name: checklistId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '201':
          description: 시작 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeLog'
  /time-logs/{id}/stop:
    post:
      tags: [Time Logs]
      summary: 타이머 정지
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 정지 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeLog'
...
  /receipts:
    get:
      tags: [Receipts]
      summary: 영수증 목록 조회 (아카이브 탭)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Receipt'

    post:
      tags: [Receipts]
      summary: 영수증 생성/갱신
      description: 새 영수증을 생성하거나, 해당 날짜의 영수증이 이미 있으면 갱신합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReceiptDto'
      responses:
        '201':
          description: 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Receipt'
        '200':
          description: 갱신 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Receipt'
  /receipts/{date}:
    get:
      tags: [Receipts]
      summary: 특정 날짜 영수증 조회
      parameters:
        - name: date
          in: path
          required: true
          description: 조회할 날짜 (YYYY-MM-DD 형식)
          schema:
            type: string
            format: date
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Receipt'
        '404':
          description: 해당 날짜의 영수증이 없습니다

    delete:
      tags: [Receipts]
      summary: 영수증 삭제
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
            format: date
      responses:
        '204':
          description: 삭제 성공

# ============================================================
# WebSocket API (Socket.IO)
# ============================================================
# OpenAPI 3.0 does not natively support WebSocket.
# Below is AsyncAPI-style documentation using vendor extensions.
# ============================================================
...